requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.11.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
defusedxml==0.8.0rc2
Deprecated==1.2.14
Django==4.2.7
django-allauth==0.58.2
django-braces==1.15.0
django-oauth-toolkit==2.3.0
django-phonenumber-field==7.2.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
ecdsa==0.18.0
idna==3.6
inflection==0.5.1
jsonschema==4.19.2
jsonschema-specifications==2023.11.1
jwcrypto==1.5.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.11.2
phonenumbers==8.13.25
Pillow==10.1.0
platformdirs==4.0.0
psycopg==3.1.12
psycopg-binary==3.1.12
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python-jose==3.3.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.12.0
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
tomli==2.0.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
wrapt==1.16.0
